feat: fix delete table service

+++ src/services/tables/deleteTable.service.ts
@@ -1,8 +1,17 @@
+import { AppError } from "./../../errors/AppError";
 import { Table } from "./../../entities/table.entity";
 import AppDataSource from "../../data-source";
 
-export const deleteTableService = async (tableId) => {
+export const deleteTableService = async (tableId): Promise<{}> => {
   const tableRepository = AppDataSource.getRepository(Table);
+
+  const findTable = tableRepository.findOneBy({
+    id: tableId,
+  });
+
+  if (!findTable) {
+    throw new AppError("Table not exist", 404);
+  }
 
   const tableDelete = await tableRepository.delete({
     id: tableId,
